cleanup and  traits + docs
