+++ xmtp_api_d14n/src/protocol/mod.rs
@@ -3,8 +3,6 @@
 pub mod traits;
 pub use traits::*;
 
-pub mod envelopes;
-
 pub mod extractors;
 pub use extractors::*;
 
@@ -12,5 +10,3 @@ mod in_memory_cursor_store;
 pub use in_memory_cursor_store::*;
 
 mod impls;
-
-mod xmtp_query;
