+++ xmtp_api_d14n/src/protocol/traits/xmtp_query.rs
@@ -1,0 +1,51 @@
+//! XmtpQuery allows accessing the network while bypassing any local cursor cache.
+use super::*;
+
+// XMTP Query queries the network for any envelopes
+/// matching the cursor criteria given.
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
+pub trait XmtpQuery: Send + Sync {
+    type Error: RetryableError + Send + Sync + 'static;
+    /// Query every [`Topic`] at [`GlobalCursor`]
+    async fn query_at(
+        &self,
+        topic: Topic,
+        at: Option<GlobalCursor>,
+    ) -> Result<XmtpEnvelope, Self::Error>;
+}
+
+// hides implementation detail of XmtpEnvelope/traits
+/// Envelopes from the XMTP Network received from a general [`XmtpQuery`]
+pub struct XmtpEnvelope {
+    inner: Box<dyn EnvelopeCollection<'static> + Send + Sync>,
+}
+
+impl XmtpEnvelope {
+    pub fn new(envelope: impl EnvelopeCollection<'static> + Send + Sync + 'static) -> Self {
+        Self {
+            inner: Box::new(envelope) as Box<_>,
+        }
+    }
+
+    pub fn len(&self) -> usize {
+        self.inner.len()
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.inner.is_empty()
+    }
+
+    pub fn group_messages(&self) -> Result<Vec<GroupMessage>, EnvelopeError> {
+        Ok(self.inner.group_messages()?.into_iter().flatten().collect())
+    }
+
+    pub fn welcome_messages(&self) -> Result<Vec<WelcomeMessage>, EnvelopeError> {
+        Ok(self
+            .inner
+            .welcome_messages()?
+            .into_iter()
+            .flatten()
+            .collect())
+    }
+}
