--- xmtp_api_d14n/src/protocol/xmtp_query.rs
@@ -1,1 +1,0 @@
-//! Generic XMTP Query Implementation
